test(BT) : 2023. 1 Unity BT

+++ 0_Scripts/Project/RoboRobo/1_Custom/After/0_Character/4_BehaviorTree/Enemy/HealthNode.cs
@@ -6,9 +6,9 @@ public class HealthNode : Node
 {
 
     private BTBoss ai;
-    private float threshold;
+    private int threshold;
 
-    public HealthNode(BTBoss ai, float threshold)
+    public HealthNode(BTBoss ai, int threshold)
     {
 
         this.ai = ai;
